+++ vite.config.js
@@ -3,6 +3,7 @@ import { svelte } from '@sveltejs/vite-plugin-svelte';
 
 export default defineConfig({
   root: 'src',
+  envDir: '..',
   plugins: [svelte()],
   build: {
     outDir: '../dist',